pyproject.toml
[tool.poetry]
name = "whoami"
version = "0.1.0"
description = ""
authors = ["Thomas R Storey <thomas@feathr.co>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pymongo = {extras = ["srv"], version = "^4.0.1"}
python-dotenv = "^0.19.2"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest = "7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
